Move MapConfigurator to maps module
