+++ collect_app/src/main/java/org/odk/collect/android/geo/GoogleMapConfigurator.java
@@ -17,6 +17,7 @@ import org.odk.collect.android.geo.MbtilesFile.LayerType;
 import org.odk.collect.android.utilities.PlayServicesChecker;
 import org.odk.collect.androidshared.ui.PrefUtils;
 import org.odk.collect.androidshared.ui.ToastUtils;
+import org.odk.collect.maps.MapConfigurator;
 import org.odk.collect.shared.settings.Settings;
 
 import java.io.File;
