+++ maps/build.gradle
@@ -54,6 +54,7 @@ dependencies {
     implementation Dependencies.kotlin_stdlib
     implementation Dependencies.androidx_fragment
     implementation Dependencies.androidx_annotations
+    implementation Dependencies.androidx_preference_ktx
 
     testImplementation Dependencies.junit
     testImplementation Dependencies.hamcrest
